+++ src/neptune_tensorflow_keras/__init__.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from neptune_tensorflow_keras.impl import (
+    NeptuneCallback,
+    __version__,
+)